# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API dokumentace aplikace",
        "contact": {
            "name": "API Support",
            "email": "info@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.example.com/v1",
            "description": "Test server"
        }
    ],
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Seznam všech knih",
                "description": "Vrátí seznam všech knih s možností filtrování, řazení a stránkování",
                "operationId": "getBooksList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Číslo stránky",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Počet položek na stránku",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sloupec pro řazení",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "title",
                                "published_at",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Směr řazení",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filtrovat podle názvu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "Filtrovat podle ID autora",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrovat podle ID kategorie",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Úspěšná operace",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Vytvoření nové knihy",
                "description": "Vytvoří novou knihu a vrátí její data",
                "operationId": "storeBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Harry Potter"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-80-00-05942-8"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Popis knihy..."
                                    },
                                    "page_count": {
                                        "type": "integer",
                                        "example": 320
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "cs"
                                    },
                                    "author_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kniha byla úspěšně vytvořena",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validační chyba"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Detail knihy",
                "description": "Vrátí detailní informace o knize podle ID",
                "operationId": "getBookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID knihy",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Úspěšná operace",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kniha nebyla nalezena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Aktualizace knihy",
                "description": "Aktualizuje knihu podle ID a vrátí její data",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID knihy",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Harry Potter (aktualizováno)"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-80-00-05942-8"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-15"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nový popis knihy..."
                                    },
                                    "page_count": {
                                        "type": "integer",
                                        "example": 320
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "cs"
                                    },
                                    "author_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3
                                        ]
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            2,
                                            4
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kniha byla úspěšně aktualizována",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kniha nebyla nalezena"
                    },
                    "422": {
                        "description": "Validační chyba"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Smazání knihy",
                "description": "Smaže knihu podle ID",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID knihy",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Kniha byla úspěšně smazána"
                    },
                    "404": {
                        "description": "Kniha nebyla nalezena"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}